move new test into its own file....at the end
