+++ test/functional/apps/management/_index_pattern_create_delete.js
@@ -17,7 +17,6 @@ export default function ({ getService, getPageObjects }) {
   const testSubjects = getService('testSubjects');
   const find = getService('find');
   const PageObjects = getPageObjects(['settings', 'common', 'header']);
-  const es = getService('es');
 
   describe('creating and deleting default index', function describeIndexTests() {
     before(async function () {
@@ -133,27 +132,6 @@ export default function ({ getService, getPageObjects }) {
           });
         });
       });
-    });
-
-    it('data view creation with exclusion', async () => {
-      await es.transport.request({
-        path: '/index-a/_doc',
-        method: 'POST',
-        body: { user: 'matt' },
-      });
-
-      await es.transport.request({
-        path: '/index-b/_doc',
-        method: 'POST',
-        body: { title: 'hello' },
-      });
-
-      await PageObjects.settings.createIndexPattern('index-*,-index-b');
-
-      const fieldCount = await PageObjects.settings.getFieldsTabCount();
-
-      // five metafields plus keyword and text version of 'user' field
-      expect(parseInt(fieldCount, 10)).to.be(7);
     });
   });
 }
